fix: conflict merge

+++ ems/src/main/java/com/group5/ems/dto/response/HrDashboardMetricsDTO.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record HrDashboardMetricsDTO(
-    int activeEmployees,
+    Long activeEmployees,
     int pendingLeaveRequests,
     int openJobPosts,
     int pendingWorkflowRequests
